ci-framework branch initial commit
As documented [1], the Validated Architecture CRs require
customization for the environment where they are utilized.

This commit [2] has values which match a lot of what an
environment deployed by the ci-framework so I am mirroring
it here so people can easily clone it. I will add follow up
patches to make it eaiser to use if you use ci-framework.

[1] https://github.com/openstack-k8s-operators/architecture/tree/main/validated_arch_1#considerations
[2] https://github.com/abays/architecture/commit/c2926af65d2099bbb710d847991030510442dec4

Signed-off-by: John Fulton <fulton@redhat.com>