Configure OVN Controller to act as gateway
To configure nicMappings(external net-attach-def) for
ovnController CR as it requires dedicated nic with
external connectivity on the worker nodes.

With network isolation we already setup an additional
nic on the worker nodes and as that is already used by
MetalLB and net-attach-definitions it cannot be
shared by OVN Controllers as it uses host-device
plugin and requires a dedicated NIC.

In order to use same interface for ovn controller,
a linux bridge is created with with ctlplane interface as one
of it's port. Utilizing this bridge a net-attach-def of
bridge type is created for usage by the OVN Controllers.

If there is additional external NIC is used in worker then
that can be used.

Issue: OSPRH-3947
(cherry picked from commit 05ea207ae6aeebc28124cb47112dd401bf3a5132)
(cherry picked from commit 1e73f23c571ab189b2a6ca8aa443ebb21f1c7369)